+++ utbot-fuzzers/src/main/kotlin/org/utbot/fuzzer/RandomExtensions.kt
@@ -39,21 +39,5 @@ fun Long.invertBit(bitIndex: Int): Long {
     return this xor (1L shl bitIndex)
 }
 
-
-fun ModelProvider.assembleModel(id: Int, constructorId: ConstructorId, params: List<FuzzedValue>): FuzzedValue {
-    val instantiationChain = mutableListOf<UtStatementModel>()
-    return UtAssembleModel(
-        id,
-        constructorId.classId,
-        "${constructorId.classId.name}${constructorId.parameters}#" + id.hex(),
-        instantiationChain = instantiationChain,
-        modificationsChain = mutableListOf()
-    ).apply {
-        instantiationChain += UtExecutableCallModel(null, constructorId, params.map { it.model }, this)
-    }.fuzzed {
-        summary = "%var% = ${constructorId.classId.simpleName}(${constructorId.parameters.joinToString { it.simpleName }})"
-    }
-}
-
 fun Int.hex(): String =
     toString(16)
