+++ utbot-fuzzers/src/main/kotlin/org/utbot/fuzzer/objects/AssembleModelUtils.kt
@@ -1,0 +1,25 @@
+package org.utbot.fuzzer.objects
+
+import org.utbot.framework.plugin.api.ConstructorId
+import org.utbot.framework.plugin.api.UtAssembleModel
+import org.utbot.framework.plugin.api.UtExecutableCallModel
+import org.utbot.framework.plugin.api.UtStatementModel
+import org.utbot.fuzzer.FuzzedValue
+import org.utbot.fuzzer.ModelProvider
+import org.utbot.fuzzer.hex
+
+
+fun ModelProvider.assembleModel(id: Int, constructorId: ConstructorId, params: List<FuzzedValue>): FuzzedValue {
+    val instantiationChain = mutableListOf<UtStatementModel>()
+    return UtAssembleModel(
+        id,
+        constructorId.classId,
+        "${constructorId.classId.name}${constructorId.parameters}#" + id.hex(),
+        instantiationChain = instantiationChain,
+        modificationsChain = mutableListOf()
+    ).apply {
+        instantiationChain += UtExecutableCallModel(null, constructorId, params.map { it.model }, this)
+    }.fuzzed {
+        summary = "%var% = ${constructorId.classId.simpleName}(${constructorId.parameters.joinToString { it.simpleName }})"
+    }
+}
